Rename java package to `io.kokuwa.maven.helm`.
