--- src/main/java/com/kiwigrid/helm/maven/plugin/pojo/RepoType.java
@@ -1,11 +1,0 @@
-package com.kiwigrid.helm.maven.plugin.pojo;
-
-/**
- * Enumeration for supported repository types
- *
- * @author Axel Köhler
- * @since 2.4
- */
-public enum RepoType {
-	CHARTMUSEUM, ARTIFACTORY, NEXUS
-}
