+++ src/main/java/io/kokuwa/maven/helm/exception/BadUploadException.java
@@ -1,0 +1,17 @@
+package io.kokuwa.maven.helm.exception;
+
+/**
+ * Indicates that the chart upload was invalid.
+ *
+ * @author Axel Koehler
+ * @since 2.1
+ */
+public class BadUploadException extends Exception {
+
+	public BadUploadException() {
+	}
+
+	public BadUploadException(String message) {
+		super(message);
+	}
+}
