+++ src/main/java/io/kokuwa/maven/helm/pojo/RepoType.java
@@ -1,0 +1,11 @@
+package io.kokuwa.maven.helm.pojo;
+
+/**
+ * Enumeration for supported repository types
+ *
+ * @author Axel Köhler
+ * @since 2.4
+ */
+public enum RepoType {
+	CHARTMUSEUM, ARTIFACTORY, NEXUS
+}
