+++ src/test/java/io/kokuwa/maven/helm/junit/SystemProperty.java
@@ -1,0 +1,26 @@
+package io.kokuwa.maven.helm.junit;
+
+import java.lang.annotation.*;
+
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target({ ElementType.METHOD, ElementType.TYPE })
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface SystemProperty {
+
+    String name();
+
+    String value() default "";
+
+    boolean unset() default false;
+
+    @Target({ ElementType.METHOD, ElementType.TYPE })
+    @Retention(RUNTIME)
+    @Documented
+    @interface SystemProperties {
+
+        SystemProperty[] value();
+
+    }
+}
