+++ src/test/resources/io/kokuwa/maven/helm/app-0.1.0.tgz
